+++ tests/test_endpoints.py
@@ -1,4 +1,6 @@
+import datetime
 from unittest import TestCase
+from unittest.mock import MagicMock
 
 from piccolo.apps.user.tables import BaseUser
 from piccolo.columns.column_types import (
@@ -8,11 +10,12 @@ from piccolo.columns.column_types import (
     Timestamp,
     Varchar,
 )
-from piccolo.table import Table
+from piccolo.table import Table, create_db_tables_sync, drop_db_tables_sync
+from piccolo_api.crud.hooks import Hook, HookType
 from piccolo_api.session_auth.tables import SessionsBase
 from starlette.testclient import TestClient
 
-from piccolo_admin.endpoints import TableConfig, get_all_tables
+from piccolo_admin.endpoints import TableConfig, create_admin, get_all_tables
 from piccolo_admin.example import APP
 from piccolo_admin.version import __VERSION__
 
@@ -139,11 +142,10 @@ class TestForms(TestCase):
     credentials = {"username": "Bob", "password": "bob123"}
 
     def setUp(self):
-        SessionsBase.create_table(if_not_exists=True).run_sync()
-        BaseUser.create_table(if_not_exists=True).run_sync()
-        BaseUser(
+        create_db_tables_sync(SessionsBase, BaseUser, if_not_exists=True)
+        BaseUser.create_user_sync(
             **self.credentials, active=True, admin=True, superuser=True
-        ).save().run_sync()
+        )
 
     def tearDown(self):
         SessionsBase.alter().drop_table().run_sync()
@@ -317,15 +319,13 @@ class TestTables(TestCase):
     credentials = {"username": "Bob", "password": "bob123"}
 
     def setUp(self):
-        SessionsBase.create_table(if_not_exists=True).run_sync()
-        BaseUser.create_table(if_not_exists=True).run_sync()
-        BaseUser(
+        create_db_tables_sync(SessionsBase, BaseUser, if_not_exists=True)
+        BaseUser.create_user_sync(
             **self.credentials, active=True, admin=True, superuser=True
-        ).save().run_sync()
+        )
 
     def tearDown(self):
-        SessionsBase.alter().drop_table().run_sync()
-        BaseUser.alter().drop_table().run_sync()
+        drop_db_tables_sync(SessionsBase, BaseUser)
 
     def test_tables(self):
         """
@@ -441,3 +441,76 @@ class TestTables(TestCase):
         self.assertEqual(
             response.json(), {"error": "Language does not exist."}
         )
+
+
+class TestHooks(TestCase):
+
+    credentials = {"username": "Bob", "password": "bob123"}
+
+    def setUp(self):
+        create_db_tables_sync(SessionsBase, BaseUser, Post, if_not_exists=True)
+        BaseUser.create_user_sync(
+            **self.credentials, active=True, admin=True, superuser=True
+        )
+
+    def tearDown(self):
+        drop_db_tables_sync(SessionsBase, BaseUser, Post)
+
+    def test_hooks(self):
+        """
+        If a hook is passed to ``TableConfig``, make sure it gets called.
+        """
+        mock = MagicMock()
+
+        def hook(row, mock: MagicMock = mock):
+            mock()
+            row.name = "New Post 1"
+            return row
+
+        app = create_admin(
+            tables=[
+                TableConfig(
+                    table_class=Post,
+                    hooks=[Hook(hook_type=HookType.pre_save, callable=hook)],
+                )
+            ]
+        )
+
+        client = TestClient(app)
+
+        # To get a CSRF cookie
+        response = client.get("/")
+        csrftoken = response.cookies["csrftoken"]
+
+        # Login
+        payload = dict(csrftoken=csrftoken, **self.credentials)
+        client.post(
+            "/auth/login/",
+            json=payload,
+            headers={"X-CSRFToken": csrftoken},
+        )
+
+        # Now try creating a new row
+        response = client.post(
+            "/api/tables/post/",
+            json={
+                "name": "New Post",
+                "content": "Some content",
+                "rating": 100,
+                "created": datetime.datetime.now().isoformat(),
+            },
+            headers={"X-CSRFToken": csrftoken},
+        )
+        self.assertEqual(response.status_code, 201)
+
+        # Make sure the row is created with the correct data (the hook modified
+        # the name attribute).
+        self.assertFalse(
+            Post.exists().where(Post.name == "New Post").run_sync()
+        )
+        self.assertTrue(
+            Post.exists().where(Post.name == "New Post 1").run_sync()
+        )
+
+        # Make sure the hook was only called once.
+        mock.assert_called_once()